+++ qbittorrentapi/log.py
@@ -1,9 +1,8 @@
 from qbittorrentapi.decorators import login_required
 from qbittorrentapi.decorators import response_json
-from qbittorrentapi.helpers import APINames
-from qbittorrentapi.helpers import ClientCache
-from qbittorrentapi.helpers import List
-from qbittorrentapi.helpers import ListEntry
+from qbittorrentapi.definitions import APINames
+from qbittorrentapi.definitions import ClientCache
+from qbittorrentapi.definitions import List, ListEntry
 from qbittorrentapi.request import Request
