+++ setup.py
@@ -11,7 +11,8 @@ setup(
     install_requires=['attrdict>=2.0.0',
                       'requests>=2.16.0',
                       'urllib3>=1.24.2',
-                      'six'],
+                      'six',
+                      'enum34'],
     url='https://github.com/rmartin16/qbittorrent-api',
     author='Russell Martin',
     description='Python client for qBittorrent v4.1+ Web API',
